play_board/requirements.txt
Django==3.2
python-dotenv==0.21.0
Pillow==9.4.0
geopy==2.3.0
sorl-thumbnail==12.9.0
requests==2.28.2
folium==0.14.0
django-debug-toolbar==3.4.0
APScheduler==3.10.0
psycopg2-binary
pyTelegramBotAPI==4.10.0
sentry-sdk
gunicorn==20.1.0